+++ src/Array.ts
@@ -737,3 +737,17 @@ export const maximum: <A>(ord: Ord<A>) => (xs: NonEmptyArray<A>) => A = flow(
   getJoinSemigroup,
   NEA.fold,
 )
+
+/**
+ * Append two arrays in terms of a semigroup. In effect, a functional wrapper
+ * around `Array.prototype.concat`.
+ *
+ * @example
+ * import { concat } from 'fp-ts-std/Array';
+ *
+ * assert.deepStrictEqual(concat([1, 2])([3, 4]), [1, 2, 3, 4]);
+ *
+ * @since 0.9.0
+ */
+export const concat = <A>(xs: Array<A>): Endomorphism<Array<A>> => ys =>
+  A.getMonoid<A>().concat(xs, ys)